refactor: import Box from index
