+++ src/components/HomePageTabs/TeachersTab/TeachersTab.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 
 import ImageContainer from "../ImageContainer/ImageContainer";
 
-import Box from "@/components/SharedComponents/Box/Box";
+import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
 import { Heading } from "@/components/Typography";
 import Typography from "@/components/Typography/Typography";
