+++ src/components/pages/CurriculumInfo/tabs/OverviewTab/OverviewTab.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-import Box from "@/components/SharedComponents/Box/Box";
+import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
 import { P, Heading, UL, LI } from "@/components/Typography";
 import Card from "@/components/Card/Card";
